Added Redis as sessions DB, fixed crash on API view in browser
After adding CORS-header settings to the project, we had a problem that we couldn't view API directly from the browser, because Django crashed due to the inability to save user session data, because we don't have SQL DB configured.
This commit fixes this problem by configuring Django to use Redis as the default DB for cache and sessions.